Add ability to change time in the program
* Remove the functionality of changing the time according to the ratio
of the radial fill. This functionality was useful when first developing
the functionality and is no longer needed.

* Add functionality to get input on the new time to set the clock too
and change the time after the user confirms the new time.

* Add DayCycle.cs update calls to UpdateTimeToGivenTime to run updates
after the new time is confirmed. This will allow the user to see
the time change.

* Add functionality to pause the game when the menu is toggled on.

* Add Green material to the plane to visually separate the plane and the
cube. This also allows us to see the shadows cast on the cube more.